backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
openai>=2.20.0
qdrant-client==1.12.1
lunar-python==1.3.6
pydantic==2.10.4
